refactoring authInfo
